s1756: refactor and add put capability, schema val off default
Use a single method in the data client called "transfer_file"
to handle both get and put.

Schema validation is slow, so turn off by default. Also make
validation a property of the TransferReader rather than an
option to the read method so that it can be initialized once,
and then re-used throughout the client.