s1756: more standardization, canfar/cadc separation, README
For now both the CADC and CANFAR clients reside in the same repo,
but clients that hit CANFAR services will go in the canfar package,
and CADC-specific clients in the cadc package. cadc is
presently the lower-level of the two (common base classes live
there). Not much motivation at this stage to have a third
generic package for base classes.

Renamed the gms* scripts to canfar-<something> to match the style
decided on for the cadc data service client (cadc-copy).

Added a fairly complete README.md

Fixed up rununittests to avoid generating coverage for
external dependencies in the virtual environment ($VIRTUAL_ENV).