s1756: initialize logger inside constructor
Having the initialization as part of a separate method called
after object instantiation meant that none of the log messages
emitted by the constructor were displayed. It also seemed silly
having all of the verbosity parameters from the parser passed
explicitly to this get_logger method. So, now there is a method
in the parser that converts --verbose, --debug etc. into a single
numerical value, get_log_level(), which is called after the
arguments have been parsed. This single value goes into the
client constructor, and now we have useful logging messages.