t69114: fix blocking behaviour on streamed gets, specify chunk_size
When getting a large file have the status check on the Response
object from the get (prior to iterating over the chunks) wound
up blocking.

Instead, perform the check after the streaming. Note that
the "response.text" will already be consumed at this point,
so we need to handle that exception in
canfar.common.client.Client.check_status().

DataClient constructor now takes chunk_size as option.